+++ plugin.server.js
@@ -1,0 +1,4 @@
+export default function (ctx, inject) {
+  // Inject winston logger instance to the Nuxt context as $winstonLog
+  ctx.$winstonLog = process.winstonLog || {}
+}